+++ src/sdk/clients/decorators/mee/getPermitQuote.test.ts
@@ -12,7 +12,10 @@ import {
   type FeeTokenInfo,
   type Instruction,
   type Trigger,
-  getFusionQuote
+  executeSignedQuote,
+  getFusionQuote,
+  signPermitQuote,
+  waitForSupertransactionReceipt
 } from "."
 import { getTestChainConfig, toNetwork } from "../../../../test/testSetup"
 import type { NetworkConfig } from "../../../../test/testUtils"
@@ -20,8 +23,15 @@ import { getBalance } from "../../../../test/testUtils"
 import type { MultichainSmartAccount } from "../../../account/toMultiChainNexusAccount"
 import { toMultichainNexusAccount } from "../../../account/toMultiChainNexusAccount"
 import { mcUSDC } from "../../../constants/tokens"
-import { runtimeERC20BalanceOf } from "../../../modules/utils/composabilityCalls"
-import { type MeeClient, createMeeClient } from "../../createMeeClient"
+import {
+  greaterThanOrEqualTo,
+  runtimeERC20BalanceOf
+} from "../../../modules/utils/composabilityCalls"
+import {
+  DEFAULT_MEE_NODE_URL,
+  type MeeClient,
+  createMeeClient
+} from "../../createMeeClient"
 import getPermitQuote from "./getPermitQuote"
 
 describe("mee.getPermitQuote", () => {
@@ -37,8 +47,6 @@ describe("mee.getPermitQuote", () => {
   let paymentChain: Chain
   let targetChain: Chain
   let transports: Transport[]
-
-  const recipient: Address = "0x3079B249DFDE4692D7844aA261f8cf7D927A0DA5"
 
   beforeAll(async () => {
     network = await toNetwork("MAINNET_FROM_ENV_VARS")
@@ -154,8 +162,8 @@ describe("mee.getPermitQuote", () => {
     const trigger: Trigger = {
       chainId: paymentChain.id,
       tokenAddress,
-      amount: totalBalance,
-      useMaxAvailableAmount: true
+      amount: 6000n,
+      includeFee: true
     }
 
     // withdraw
@@ -187,6 +195,45 @@ describe("mee.getPermitQuote", () => {
     expect(fusionQuote.trigger.amount).toBeGreaterThan(0n)
   })
 
+  // This test uses available usdc on the eoa on mainnet, so should be skipped
+  test.skip("should demo behaviour of max available amount", async () => {
+    const vitalik = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
+    const chainId = paymentChain.id
+    const mcNexusAddress = mcNexus.addressOn(paymentChain.id, true)
+    const trigger: Trigger = {
+      chainId,
+      tokenAddress,
+      amount: 6000n,
+      includeFee: true
+    }
+
+    const transferInstruction = await mcNexus.buildComposable({
+      type: "transfer",
+      data: {
+        chainId,
+        tokenAddress,
+        recipient: vitalik,
+        amount: runtimeERC20BalanceOf({
+          targetAddress: mcNexusAddress,
+          tokenAddress,
+          constraints: [greaterThanOrEqualTo(1n)]
+        })
+      }
+    })
+
+    const fusionQuote = await meeClient.getPermitQuote({
+      trigger,
+      instructions: [transferInstruction], // inx 1 => transferFrom (Runtime) + Dev userOps
+      feeToken
+    })
+
+    const signedQuote = await signPermitQuote(meeClient, { fusionQuote }) // Permit with 20k
+    const { hash } = await executeSignedQuote(meeClient, { signedQuote })
+
+    const receipt = await waitForSupertransactionReceipt(meeClient, { hash })
+    expect(receipt.transactionStatus).toBe("MINED_SUCCESS")
+  })
+
   test("should add gas fees to amount when not using max available amount", async () => {
     const client = createPublicClient({
       chain: paymentChain,
@@ -197,8 +244,8 @@ describe("mee.getPermitQuote", () => {
     const trigger: Trigger = {
       chainId: paymentChain.id,
       tokenAddress,
-      amount: amount
-      // useMaxAvailableAmount not set, should default to false
+      amount
+      // max not set, should default to false
     }
 
     // withdraw