+++ src/sdk/clients/decorators/mee/getPermitQuote.ts
@@ -1,6 +1,10 @@
 import type { BuildInstructionTypes } from "../../../account"
 import { batchInstructions } from "../../../account/utils/batchInstructions"
 import { resolveInstructions } from "../../../account/utils/resolveInstructions"
+import {
+  greaterThanOrEqualTo,
+  runtimeERC20AllowanceOf
+} from "../../../modules/utils/composabilityCalls"
 import type { BaseMeeClient } from "../../createMeeClient"
 import { type GetQuotePayload, getQuote } from "./getQuote"
 import type { GetQuoteParams } from "./getQuote"
@@ -80,21 +84,32 @@ export const getPermitQuote = async (
 
   const sender = account_.signer.address
   const recipient = account_.addressOn(trigger.chainId, true)
-
   const resolvedInstructions = await resolveInstructions(instructions)
 
   const isComposable = resolvedInstructions.some(
     ({ isComposable }) => isComposable
   )
 
+  const transferFromAmount = trigger.includeFee
+    ? runtimeERC20AllowanceOf({
+        owner: sender,
+        spender: recipient,
+        tokenAddress: trigger.tokenAddress,
+        constraints: [greaterThanOrEqualTo(1n)]
+      })
+    : trigger.amount
+
   const params: BuildInstructionTypes = {
     type: "transferFrom",
-    data: { ...trigger, recipient, sender }
+    data: {
+      tokenAddress: trigger.tokenAddress,
+      chainId: trigger.chainId,
+      amount: transferFromAmount,
+      recipient,
+      sender
+    }
   }
 
-  // The trigger transfer is the first instruction in the array.
-  // It draws funds from the eoa to the nexus account with a transferFrom call.
-  // If the instructions are composable, we build the composable transaction
   const triggerTransfer = await (isComposable
     ? account_.buildComposable(params)
     : account_.build(params))
@@ -112,15 +127,18 @@ export const getPermitQuote = async (
     ...rest
   })
 
-  // This trigger should have an amount that is the amount user wishes to spend, plus the gas fees
-  const trigger_ = {
-    ...trigger,
-    amount: trigger.useMaxAvailableAmount
-      ? BigInt(trigger.amount)
-      : BigInt(trigger.amount) + BigInt(quote.paymentInfo.tokenWeiAmount)
-  }
+  const amount = trigger.includeFee
+    ? BigInt(trigger.amount)
+    : BigInt(trigger.amount) + BigInt(quote.paymentInfo.tokenWeiAmount)
 
-  return { quote, trigger: trigger_ }
+  return {
+    quote,
+    trigger: {
+      tokenAddress: trigger.tokenAddress,
+      chainId: trigger.chainId,
+      amount
+    }
+  }
 }
 
 export default getPermitQuote